# Remove query_plan from RunningQueries proc output, changed to lazy load query plan on demand