--- a/DBADashDB/DBADashDB.sqlproj
+++ b/DBADashDB/DBADashDB.sqlproj
@@ -824,6 +824,7 @@
     <Build Include="dbo\Stored Procedures\RepositoryMetricsConfig_Upd.sql" />
     <Build Include="dbo\Stored Procedures\RepositoryMetricsConfig_Del.sql" />
     <Build Include="dbo\Stored Procedures\DatabaseFinder_Get.sql" />
+    <Build Include="dbo\Stored Procedures\QueryPlan_Get.sql" />
   </ItemGroup>
   <ItemGroup>
     <PostDeploy Include="Script.PostDeployment1.sql" />