Simplified PR of existing previous PR: https://github.com/trimble-oss/dba-dash/pull/1165

I removed the code which compresses large text fields and only included the changes to lazy load query plans on demand from the GUI.

If a query plan exists (`has_plan = true`) then it will attempt to load it using the new proc `QueryPlan_Get`. Once the plan has been loaded once for that row, it will retain it in case it is requested again.

Further optimizations _could_ be done to cache the plan for all other sessions with the same plan, but for now this seems good enough.

One arguable decision of note that was made was adding `query_plan_hash_bin` and `plan_handle_bin` to the output of `RunningQueries_Get`. These were added to be used for performing lookups via the new `QueryPlan_Get` proc.

It could be argued this is unnecessary and it would be easier to rely on the existing `query_plan_hash` and `plan_handle` string columns which are already part of the `RunningQueries_Get` dataset. I'm fine either way and open to changing it. The only reason I went this route was so that `QueryPlan_Get` would have more sensible datatypes for its parameters. Adding the two additional `*_bin` columns adds an extra 72 bytes per row, which I did not find to make much of a difference with vs without.

Testing was performed with various sized batches ranging from 50 sessions to ~2000. Performance improvements listed in the below table. Before / after measurements are an average of 5 tests.

| avg sessions | before (ms) | after (ms) |
|--------------|-------------|------------|
|           50 |         744 |        327 |
|          100 |       1,261 |        240 |
|          200 |       2,075 |        787 |
|          500 |      25,492 |      1,620 |
|          700 |      48,808 |      4,545 |
|        1,000 |      61,086 |      6,010 |
|        1,500 |      64,517 |      7,153 |
|        2,000 |     105,141 |      9,801 |