@DavidWiseman 2025-04-30
Thank you for the updated PR.  It works great! 👍

I think it's OK to add the query_plan_hash_bin and plan_handle_bin columns.  The string columns could have been converted in .NET to keep the data types matching in QueryPlan_Get.  

It might also be better to remove the string versions of the columns, replacing them with the binary.  By default, the DataGridView in .NET doesn't display the binary data correctly.  I'm now using a custom version of the DataGridView that automatically handles this.  There are still some issues with this though (Filter by value, export to excel) that would need to be looked at.  Something that could be improved at a later date.